Refactor AiServiceTokenStream constructor with pilot way for encapsulating parameters (#2844)
This is somewhat of a POC for how we may in the future encapsulate
classes with large numbers of parameters in their constructors.

This would allow it to be backwards-compatible and non-breaking whenever
parameters to a constructor on public-facing objects are created.

(cherry picked from commit 2e407503c8a516a9bcfa44442b070e441c850f10)